Create LatentClassAnalysisFTAF.R
This is the R file used to generate the analysis of the datasets Data_Humans.csv and Data_Animals.csv, and contains the JAGS models and the code to produce the figures. When running the file, please ensure that the plotting window is very large if you run this on a small-screened device.